refactor: share RayIntersectionT and eliminate GzRay copy
- Add include/gz/physics/RayIntersection.hh with a policy-templated
  RayIntersectionT shared between single-ray and batch features
- Replace inline struct definitions in GetRayIntersection.hh and
  GetBatchRayIntersection.hh with template aliases to the shared type
- Make GzRay a type alias for RayT<FeaturePolicy3d> (same as GzRayResult),
  eliminating the origin/target → from/to copy loop in
  SimulationFeatures::GetBatchRayIntersectionFromLastStep

Generated-By: Claude Sonnet 4.6 <noreply@anthropic.com>
